finished structure for test & train , working on output analysis

+++ src/main_opt.py
@@ -1,11 +1,16 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 import glob
+import os
 import json
-
 import pandas
+import numpy as np
+from multiprocessing import Process
+
+import pandas as pd
 
 from src import strategies
+# import strategies
 import backtrader as bt  # Import the backtrader platform
 import datetime  # For datetime objects
 # import os.path  # To manage paths
@@ -30,6 +35,18 @@ TFRAMES = dict(
     years=bt.TimeFrame.Years)
 
 
+class NpEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, np.integer):
+            return int(obj)
+        elif isinstance(obj, np.floating):
+            return float(obj)
+        elif isinstance(obj, np.ndarray):
+            return obj.tolist()
+        else:
+            return super(NpEncoder, self).default(obj)
+
+
 def runstrat_opt(settings, **kwargs):
     # clock the start of the process
     tstart = process_time()
@@ -51,13 +68,16 @@ def runstrat_opt(settings, **kwargs):
         strategies.OptStrategy,
         **kwargs["signal"].get(signal),
         signal=signal,
-        # printlog=False
+        printlog=False
         )
 
+    # Get a data source path
     datapath = args.data
+
+    # Pass it to the backtrader datafeed and add it to the cerebro
     data = bt.feeds.PandasData(dataname=pandasdatafeed(datapath, args=args),
-                               fromdate=args.fromdate,
-                               todate=args.todate)
+                               fromdate=args.train["fromdate"],      # fromdate=args.fromdate,
+                               todate=args.train["todate"])          # todate=args.todate)
     cerebro.adddata(data)
 
     # Set our desired cash start
@@ -100,20 +120,20 @@ def analyzers_log(settings, results):
             params = strat.params.__dict__
             [params.pop(param) for param in params_pop if param in params]
             analysis_value = {analyzer.__class__.__name__: analyzer.get_analysis() for analyzer in strat.analyzers}
-            analysis_key = json.dumps(params)
+            analysis_key = json.dumps(params, cls=NpEncoder)
             analyzers_dict.update({analysis_key: analysis_value})
 
     filepath = settings["opt_analyzer"]["path_log"].format(params['signal'])
-    # filepath = f"./data/analyzers_opt/analyzer_{params['signal']}.json"
     json.dump(analyzers_dict, open(filepath, "w"), sort_keys=True, indent=4)
     return None
 
 
 def analyzers_read(settings, **kwargs):
-    # filepath = "./data/analyzers_opt/analyzer_*.json"
     signal_opt = list(kwargs["signal"].keys())[0]
-    filepath = settings["opt_analyzer"]["path_log"].format(signal_opt)
+    objective_opt = settings["opt_analyzer"]["analyzer_opt"]
 
+    output = {}
+    output.update(kwargs)
     analyzers_keys = {
         "SQN": ["sqn"],
         "VWR": ["vwr"],
@@ -121,38 +141,38 @@ def analyzers_read(settings, **kwargs):
         # "TimeDrawDown": ["maxdrawdown", "maxdrawdownperiod"]
     }
 
-    output = {
-        "fromdate": kwargs.get("fromdate"),
-        "todate": kwargs.get("todate"),
-        "signal": {}
-    }
+    filepath = settings["opt_analyzer"]["path_log"].format(signal_opt)
+    analyzers = json.load(open(filepath, 'r'))
+    analyzers_df = pandas.DataFrame()
+    df = pandas.DataFrame(analyzers).T
 
-    files = glob.glob(filepath)
-    for file in files:
-        analyzers = json.load(open(file, 'r'))
-        analyzers = {tuple(json.loads(key).items()): value for key, value in analyzers.items()}
-        analyzers_df = pandas.DataFrame()
-        df = pandas.DataFrame(analyzers).T
-        for key, values in analyzers_keys.items():
-            df_sample = df[key].apply(pandas.Series).loc[:, values]
-            analyzers_df = pandas.concat([analyzers_df, df_sample], axis=1)
+    # filter params of analyzers
+    for key, values in analyzers_keys.items():
+        df_sample = df[key].apply(pandas.Series).loc[:, values]
+        analyzers_df = pandas.concat([analyzers_df, df_sample], axis=1)
 
-        objective_opt = settings["opt_analyzer"]["analyzer_opt"]
-        best = analyzers_df.sort_values(by=objective_opt, ascending=False).head(1)
-        best_index = dict(best.first_valid_index())
-        signal = best_index.pop("signal")
-        output["signal"][signal] = best_index
-        output["signal"][signal][objective_opt] = best[objective_opt].values[0]
-        output["signal"][signal]["sqn"] = best["sqn"].values[0]
+    best = analyzers_df.sort_values(by=objective_opt, ascending=False).head(1)
+    best_index = json.loads(best.first_valid_index())
+    signal = best_index.pop("signal")
+    output["signal"][signal] = best_index
+
+    output["signal"][signal]["analyzer_opt"] = {}
+    output["signal"][signal]["analyzer_opt"][objective_opt] = best[objective_opt].values[0]
+    # prototipe for multiobjective
+    output["signal"][signal]["analyzer_opt"]["sqn"] = best["sqn"].values[0]
+    # output["signal"][signal][objective_opt] = best[objective_opt].values[0]
+    # output["signal"][signal]["sqn"] = best["sqn"].values[0]
 
     filename = settings["opt_analyzer"]["path_opt_parms"]
     with open(filename, "r+") as file:
         data = json.load(file)
         match_date = False
         for idx, row in enumerate(data):
-            if (row["fromdate"] == str(output["fromdate"])) & (row["todate"] == str(output["todate"])):
+            if (row.get("train").get("fromdate") == str(output.get("train").get("fromdate"))) &\
+                    (row.get("train").get("todate") == str(output.get("train").get("todate"))):
                 data[idx]["signal"][signal] = output["signal"][signal]
                 match_date = True
+                break
         if match_date == False:
             data.append(output)
         file.seek(0)
@@ -164,6 +184,8 @@ def analyzers_read(settings, **kwargs):
 
 def daterange_opt(settings, **kwargs):
     datapath = settings.get("opt_analyzer").get("datapath")
+    dates_file_str = pandas.read_csv(datapath)["date"].unique()
+    dates_file = [datetime.datetime.strptime(dt, "%Y.%m.%d") for dt in dates_file_str]
 
     fromdate = settings.get("opt_analyzer").get("fromdate")
     todate = settings.get("opt_analyzer").get("todate")
@@ -171,20 +193,14 @@ def daterange_opt(settings, **kwargs):
     daterange_opt_train = settings.get("opt_analyzer").get("daterange_opt_train")
 
     if (fromdate == "") & (todate == ""):
-        fromdate = datapath.split("_")[-3]
-        fromdate = datetime.datetime.strptime(fromdate, "%Y.%m.%d")
-        todate = datapath.split("_")[-2]
-        todate = datetime.datetime.strptime(todate, "%Y.%m.%d")
+        fromdate = dates_file[0]
+        todate = dates_file[-1]
     elif (fromdate != "") & (todate == ""):
         fromdate = datetime.datetime.strptime(fromdate, "%Y.%m.%d")
-        # todate = fromdate + datetime.timedelta(days=daterange_opt)
-        todate = datapath.split("_")[-2]
-        todate = datetime.datetime.strptime(todate, "%Y.%m.%d")
+        todate = dates_file[-1]
     elif (fromdate == "") & (todate != ""):
         todate = datetime.datetime.strptime(todate, "%Y.%m.%d")
-        # fromdate = todate - datetime.timedelta(days=daterange_opt)
-        fromdate = datapath.split("_")[-3]
-        fromdate = datetime.datetime.strptime(fromdate, "%Y.%m.%d")
+        fromdate = dates_file[0]
 
     days_train = round(daterange_opt * daterange_opt_train)
     days_test = daterange_opt - days_train
@@ -192,27 +208,78 @@ def daterange_opt(settings, **kwargs):
     dates = {}
     i = 0
 
-    while todate > last_date_bin + datetime.timedelta(days=daterange_opt):
+    # while todate > last_date_bin + datetime.timedelta(days=daterange_opt):
+    while dates_file.index(todate) > dates_file.index(last_date_bin) + daterange_opt:
         fromdate_bin_train = last_date_bin
-        todate_bin_train = fromdate_bin_train + datetime.timedelta(days=days_train)
+        # todate_bin_train = fromdate_bin_train + datetime.timedelta(days=days_train)
+        todate_bin_train_idx = dates_file.index(fromdate_bin_train) + days_train - 1
+        todate_bin_train = dates_file[todate_bin_train_idx]
 
-        fromdate_bin_test = todate_bin_train + datetime.timedelta(days=1)
-        todate_bin_test = fromdate_bin_test + datetime.timedelta(days=days_test)
+        # fromdate_bin_test = todate_bin_train + datetime.timedelta(days=1)
+        fromdate_bin_test = dates_file[todate_bin_train_idx + 1]
+        todate_bin_test_idx = dates_file.index(fromdate_bin_test) + days_test - 1
+        todate_bin_test = dates_file[todate_bin_test_idx]
+        # todate_bin_test = fromdate_bin_test + datetime.timedelta(days=days_test)
 
-        last_date_bin = todate_bin_test + datetime.timedelta(days=1)
-        bin_dates = {"train": [fromdate_bin_train, todate_bin_train],
-                     "test": [fromdate_bin_test, todate_bin_test]}
+        # last_date_bin = todate_bin_test + datetime.timedelta(days=1)
+        last_date_bin = dates_file[todate_bin_test_idx + 1]
+        bin_dates = {"train": {"fromdate": fromdate_bin_train, "todate": todate_bin_train},
+                     "test": {"fromdate": fromdate_bin_test, "todate": todate_bin_test}}
         dates.update({i: bin_dates})
         i += 1
 
     return dates
 
 
-def main_opt(settings, **kwargs):
+def params_ops_signals(settings, **kwargs):
     params_settings = settings["opt_params"]
-    # params_opt = settings["opt_analyzer"]
+    signals = kwargs.get("signals")
+    dates = daterange_opt(settings)
+    params_opt = []
+
+    # update params_settings to range
+    for signal, values in params_settings.items():
+        for param, param_value in values.items():
+            if len(param_value) > 1:
+                params_settings[signal][param] = np.arange(*param_value)
+            else:
+                params_settings[signal][param] = param_value
+
+    # update params for each date and signal
+    for params in dates.values():
+        for signal in signals:
+
+            # Filter Signal to be used.
+            params_signal = {key: params_settings.get(key) for key in params_settings if key in signal}
+            params.update({"signal": params_signal})
+            params_opt.append(params.copy())
+
+    return params_opt
+
+
+def params_ops_validate(settings, **kwargs):
+    output = {}
+    output.update(kwargs)
+    signal_opt = list(kwargs["signal"].keys())[0]
     filename = settings["opt_analyzer"]["path_opt_parms"]
-    json.dump([], open(filename, "w"), sort_keys=True, indent=4)
+
+    with open(filename, "r") as file:
+        data = json.load(file)
+        for idx, row in enumerate(data):
+            if (row.get("train").get("fromdate") == str(output.get("train").get("fromdate"))) & \
+                (row.get("train").get("todate") == str(output.get("train").get("todate"))) & \
+                (signal_opt in list(row.get("signal").keys())):
+                file.close()
+                return True
+    return False
+
+
+def main(settings, **kwargs):
+    filename = settings["opt_analyzer"]["path_opt_parms"]
+
+    # create file to store the output if there is none
+    if os.path.isfile(filename) is False:
+        json.dump([], open(filename, "w"), sort_keys=True, indent=4)
 
     # TODO: FIX SIGNALS OF: OBVSignal; EFISignal
     # params_signal = [params.get(key) for key in signals]
@@ -220,36 +287,39 @@ def main_opt(settings, **kwargs):
     # signals = ["SMASignal"]
     # signals = ["RSISignal"]
     # signals = ["SMASignal", "RSISignal"]
-    signals = ["MACDSignal", "RSISignal", "WillRSignal", "ADXSignal", "StochasticSignal"]
-    dates = daterange_opt(settings)
+    # signals = ["MACDSignal", "RSISignal"]
+    # signals = ["MACDSignal", "RSISignal", "WillRSignal", "ADXSignal", "StochasticSignal"]
+    signals = ["MACDSignal", "RSISignal", "ADXSignal", "ElderForceIndexSignal"]
 
-    # allowed_keys = {'fromdate', 'todate'}
-    # params.__dict__.update((k, v) for k, v in kwargs.items() if k in allowed_keys)
+    params = params_ops_signals(settings, signals=signals)
 
-    for signal in signals:
-        # Filter Signal to be used.
-        params_signal = {key: params_settings.get(key) for key in params_settings if key in signal}
+    for params_opt in params:
 
-        # Fill range values.
-        params_signal = {signal: {key: range(*params_settings.get(signal).get(key))
-                                  for key in params_settings.get(signal)}
-                         for signal in params_signal}
+        # validate if params already calculated, very helpful is case the code breaks
+        if params_ops_validate(settings, **params_opt):
+            continue
 
-        params = {"signal": params_signal}
-
-        for idx, date in dates.items():
-            # TODO: ADICIONAR DATA DE TESTES PARA EVITAR DE CHAMR A FUNÇAO DATERANGE_OPT DEPOIS
-            fromdate, todate = date["train"]
-            params.update({"fromdate": fromdate, "todate": todate})
+        else:
             # run optimization strategy
-            runstrat_opt(settings, **params)
+            runstrat_opt(settings, **params_opt)
+
             # select best parameters for given signal and save
-            analyzers_read(settings, **params)
+            analyzers_read(settings, **params_opt)
+
     return None
 
 
 if __name__ == '__main__':
     settings = json.load(open("./src/settings.json"))
 
-    # runstrat()
-    main_opt(settings)
+    # clock the start of the process
+    tstart = process_time()
+
+    # main script
+    main(settings)
+
+    # clock the end of the process
+    tend = process_time()
+
+    # print out the result
+    print('Total Time used:', str(tend - tstart))
